Translate name 'scGraphne' to 'scBiG'

+++ reproducibility/clustering/run10_dca.py
@@ -1,11 +1,14 @@
+import os
+
+import h5py
 import numpy as np
 import scanpy as sc
 from dca.api import dca
-import h5py
-from scgraphne.utils import read_data,setup_seed,sample,louvain,calculate_metric
-import os
 
-for dataset in ['10X_PBMC','mouse_bladder_cell','mouse_ES_cell','human_kidney_counts','Adam','Human_pancreatic_islets','Macosko_mouse_retina']:
+from scbig.utils import read_data, setup_seed, sample, louvain, calculate_metric
+
+for dataset in ['10X_PBMC', 'mouse_bladder_cell', 'mouse_ES_cell', 'human_kidney_counts', 'Adam',
+                'Human_pancreatic_islets', 'Macosko_mouse_retina']:
     print('----------------real data: {} ----------------- '.format(dataset))
     setup_seed(0)
     method = 'DCA'
@@ -34,7 +37,7 @@ for dataset in ['10X_PBMC','mouse_bladder_cell','mouse_ES_cell','human_kidney_co
         ##sample
         seed = 10 * t
         X, Y = sample(X0, Y0, seed)
-        adata = sc.AnnData(X)
+        adata = sc.AnnData(X.astype('float'))
         print(adata)
         sc.pp.filter_genes(adata, min_cells=1)
         print("Sparsity: ", np.where(adata.X == 0)[0].shape[0] / (adata.X.shape[0] * adata.X.shape[1]))
@@ -43,7 +46,7 @@ for dataset in ['10X_PBMC','mouse_bladder_cell','mouse_ES_cell','human_kidney_co
         print(adata)
 
         ##DCA training
-        dca(adata, mode='latent',ae_type='zinb',threads=1)
+        dca(adata, mode='latent', ae_type='zinb', threads=1)
         print(adata)
 
         # louvain
@@ -53,7 +56,7 @@ for dataset in ['10X_PBMC','mouse_bladder_cell','mouse_ES_cell','human_kidney_co
         nmi_l, ari_l = np.round(calculate_metric(Y, y_pred_l), 4)
         print('Clustering Louvain: NMI= %.4f, ARI= %.4f' % (nmi_l, ari_l))
 
-        NMI_l.append(nmi_l), ARI_l.append(ari_l),N.append(n_pred)
+        NMI_l.append(nmi_l), ARI_l.append(ari_l), N.append(n_pred)
 
     np.savez(os.path.join(dir0, "results/clustering/{}/result_{}_{}.npz".format(dataset, dataset, method)),
              aril=ARI_l, nmil=NMI_l)
@@ -61,6 +64,3 @@ for dataset in ['10X_PBMC','mouse_bladder_cell','mouse_ES_cell','human_kidney_co
     print(NMI_l)
     print(ARI_l)
     print(N)
-
-
-
